Make je_malloc_conf weak on MSVC
Previously, the malloc_conf attribute wasn't as "weak" on MSVC as on
other toolchain. While it is possible to override it in an OBJ file -
in MSVC, every symbol coming from a LIB file can be overriden from an
OBJ file - it isn't possible to override it from another lib file. This
makes it impossible to override it from Rust, for instance, as it always
generates libs and links them together, instead of linking the obj files
directly.

While the weak attribute is not supported by MSVC, it can be emulated
using the /alternatename linker flag. This flag takes two symbols, and
aliases the first symbol to the second if it isn't defined during
linking. This ensures that the default value provided by jemalloc can be
overridden from all contexts, whether the override comes from an obj or
a lib, matching the behavior of gcc's weak attribute.

This is supported by MSVC link.exe since time immemorial, and by
lld-link since 3.7.0.